[#43475165] add Container/ContainerElement mixin to sequel
Refactored common code from plate and gel persister
from the sequel namespace to mixin
and include them into the relevant classes.
